fix annotation int

+++ library/src/main/kotlin/com/hansoolabs/and/view/VerticalScrollingBehavior.kt
@@ -23,7 +23,7 @@ abstract class VerticalScrollingBehavior<V : View> : CoordinatorLayout.Behavior<
         const val SCROLL_NONE = 0
     }
 
-    @IntDef(SCROLL_DIRECTION_UP.toLong(), SCROLL_DIRECTION_DOWN.toLong())
+    @IntDef(SCROLL_DIRECTION_UP, SCROLL_DIRECTION_DOWN)
     @Retention(AnnotationRetention.SOURCE)
     annotation class ScrollDirection
